+++ dialogs/pointer_info_dlg.h
@@ -17,7 +17,7 @@ public:
 
 public slots:
     void update(const glm::vec2 & point,
-                float altitude, float slope, int water_height, int soil_infiltration_rate, int soil_humidity,
+                float altitude, float slope, float water_height, int soil_infiltration_rate, int soil_humidity,
                 bool shade_set, bool shaded,
                 bool temp_set, float temp,
                 bool daily_illumination_set, int min_daily_illumination, int max_daily_illumination);
@@ -31,7 +31,7 @@ private slots:
     void setShaded(bool shaded);
     void setMinDailyIllumination(int min_daily_illumination);
     void setMaxDailyIllumination(int max_daily_illumination);
-    void setWaterHeight(int water_height);
+    void setWaterHeight(float water_height);
     void setSoilInfiltrationRate(int soil_infiltration_rate);
     void setSoilHumidity(int soil_humidity);
     void setAggregateHeight(int height);
